Sort of silly
A process that is creating the list will need to implement the shut down
handler. All in all, I think this is going against the simplicity
tenant... but it is not a simple problem. Returning to the one thread
per agent model may still suffice, but a socket should be created for
each connection attempt. The context can still be shared, however.