TOTALLY BROKEN COMMIT!!!!
New arch:

The active inventory collection is going away. The data contained within
the active collection will now be stored under the devices inventory
document. The agent will no longer directly communicate with the
inventory service. Instead, the backend service will act as a gateway
between the agent and the inventory service.

When a device is registered, it will send it's inspection payload along
with it's agent rpc information. Before transmitting this payload to the
inventory service, the backend service will update the payload with
information about itself so that the front end can relate job and task
requests to the correct backend instance:

                     |-----------|
                     | Inventory |
                     |-----------|
                       |   |   |
                      /    |    \
                    (BE1) (BE2) (BE3)
                    /
                a1..aN    a1..aN...

Jobs and Tasks collections will still exist horizontal to the BE
instances.