WIP:
Not the right way to go about this. (besides closing sockets when
possible)

Resending after re-establishing the socket results in duplicate messages
making it to the server. I need to 1) find out how to keep my sockets
alive (SO_KEEPALIVE) 2) Figure out why zmq is unable to detect a hung
socket on send.